DBM without orthogonalization
* remove the orthogonalization functionality. Orthogonalization not
useful, it will lead to overfitting.
